Create separate formUtils file
