+++ app/routes/setup-page.tsx
@@ -6,6 +6,7 @@ import {
 } from "@remix-run/react";
 import { page } from "~/store/page.client";
 import { ImageUploadField } from "~/components/ImageUploadField";
+import { getStringFromFormData } from "~/utils/formUtils";
 
 export async function clientAction({ request }: ClientActionFunctionArgs) {
   const body = await request.formData();
@@ -17,17 +18,6 @@ export async function clientAction({ request }: ClientActionFunctionArgs) {
   };
   page.set(pageData);
   return redirect("/requirements");
-}
-
-function getStringFromFormData(
-  body: FormData,
-  field: string
-): string | undefined {
-  const value = body.get(field);
-  if (typeof value === "string") {
-    return value;
-  }
-  // If the value was `null` or a `File` don't return anything
 }
 
 /** The client side loader, which runs after hydrate. Data from the serverLoader (`loader()`) is also available. */
@@ -76,7 +66,6 @@ export default function SetupPage() {
         >
           Submit
         </button>
-        {/*On success should redirect...*/}
       </Form>
     </>
   );
